Delete parts once all processed
