+++ common/src/main/scala/com/gu/media/aws/UploadAccess.scala
@@ -4,7 +4,7 @@ import com.amazonaws.auth.{AWSStaticCredentialsProvider, BasicAWSCredentials}
 import com.amazonaws.services.securitytoken.AWSSecurityTokenServiceClient
 import com.gu.media.Settings
 
-trait UploadAccess { this: Settings with AwsAccess with DynamoAccess with KinesisAccess =>
+trait UploadAccess { this: Settings with AwsAccess =>
   val userUploadBucket: String = getMandatoryString("aws.upload.bucket")
   val userUploadFolder: String = getMandatoryString("aws.upload.folder")
   val userUploadRole: String = getMandatoryString("aws.upload.role")
