+++ common/src/main/scala/com/gu/media/upload/UploadPart.scala
@@ -3,7 +3,7 @@ package com.gu.media.upload
 import org.cvogt.play.json.Jsonx
 import play.api.libs.json.Format
 
-case class UploadPart(key: String, start: Long, end: Long, uploadedToS3: Boolean = false, uploadedToYouTube: Long = 0)
+case class UploadPart(key: String, start: Long, end: Long, uploadedToS3: Boolean = false)
 
 object UploadPart {
   implicit val format: Format[UploadPart] = Jsonx.formatCaseClass[UploadPart]
